requirements.txt
flask
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fpdf
gradio
pytz
pymongo
dnspython
gunicorn
